+++ src/textual/widgets/_loading_indicator.py
@@ -54,7 +54,7 @@ class LoadingIndicator(Widget):
         self.auto_refresh = 1 / 16
 
     def render(self) -> RenderableType:
-        if self.app.animation_level is "none":
+        if self.app.animation_level == "none":
             return Text("Loading...")
 
         elapsed = time() - self._start_time